+++ src/app/components/pages/local-maps/local-maps.module.ts
@@ -3,12 +3,20 @@ import { NgModule } from '@angular/core';
 import { ButtonModule } from 'primeng/button';
 import { SharedAppModule } from '../../../shared/shared-app-module';
 import { LocalSongFilterModule } from '../../modules/local-song-filter/local-song-filter.module';
+import { PaginatorModule } from '../../modules/paginator/paginator.module';
 import { SongCardModule } from '../../modules/song-card/song-card.module';
 import { LocalMapsComponent } from './local-maps.component';
 
 @NgModule({
     declarations: [LocalMapsComponent],
-    imports: [CommonModule, SharedAppModule, ButtonModule, SongCardModule, LocalSongFilterModule],
+    imports: [
+        CommonModule,
+        SharedAppModule,
+        ButtonModule,
+        SongCardModule,
+        LocalSongFilterModule,
+        PaginatorModule
+    ],
     exports: [LocalMapsComponent],
     providers: []
 })