+++ src/app/components/pages/mapper-detail/mapper-detail.module.ts
@@ -1,23 +1,13 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
 import { ButtonModule } from 'primeng/button';
-import { InputTextModule } from 'primeng/inputtext';
-import { PaginatorModule } from 'primeng/paginator';
 import { SharedAppModule } from '../../../shared/shared-app-module';
+import { PaginatorModule } from '../../modules/paginator/paginator.module';
 import { SongCardModule } from '../../modules/song-card/song-card.module';
 import { MapperDetailComponent } from './mapper-detail.component';
 @NgModule({
     declarations: [MapperDetailComponent],
-    imports: [
-        CommonModule,
-        SongCardModule,
-        SharedAppModule,
-        ButtonModule,
-        PaginatorModule,
-        RouterModule,
-        InputTextModule
-    ],
+    imports: [CommonModule, SongCardModule, SharedAppModule, PaginatorModule, ButtonModule],
     exports: [MapperDetailComponent],
     providers: []
 })